requirements.txt
# Core Dependencies - All Explicit, No Optional

# Google Gemini (CRITICAL: Image generation model)
google-genai>=0.2.0

# Image Processing
pillow>=10.2.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# HTTP Client
requests>=2.31.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0

# Logging
loguru>=0.7.0
